# Add support for multiple attestations at once