We should bikeshed exact format and names here.

Combined with: https://github.com/chainguard-dev/terraform-publisher-apko/pull/18

I instrumented `terraform` output how long each `WriteState` invocation took, then ran a few things locally before and after.

```
terraform apply -auto-approve -target=module.go
```

With just a single module, things look not that different...

Total `WriteState` latency: 5.7s -> 4.185s (~27% drop)

Number of `WriteState` calls: 344 ->  260 (~24% drop)

But if we pull in some more modules:

```
terraform apply -auto-approve -target=module.go -target=module.dotnet -target=module.python -target=module.python-fips -target=module.go-fips
```

Total `WriteState` latency: 50s -> 33s (~34% drop)

Number of `WriteState` calls: 1078 ->  796 (~26% drop)

If we look plot each point (milliseconds), we see something like this:

![image](https://github.com/chainguard-dev/terraform-provider-cosign/assets/17863526/349a7e87-bed2-4a1e-af41-d3879c231329)

The latency for each `WriteState` grows ~linearly with the the size of the `terraform.tfstate` file, which is proportional to the number of resources, so the overall latency will grow quadratically.

Terraform also allocates two new buffer that end up growing to be the size of the `terraform.tfstate` file (and the previous file), so this also starts to cause a lot of GC pressure as the size increases.