@mattmoor 2023-09-19
I think we're leaving ~3x on the table by not enabling attesting to multiple images simultaneously.

This is still going to be ~3x the `WriteState` volume as `cosign_sign` simply because we can't recursively attest (because we want to specialize the attestation body per architecture).

I'd be fine with adding this, but similar to wanting a form of `oci_tag` that accepts `digest -> []tag`, I think we want something like `digest -> []statement` here.