@mattmoor 2023-09-20
I think I'd call the above resources something different `cosign_multi_attest` or `oci_multi_tag`?

So I'd be fine with landing this as-is and trying to measure the impact of this on our mega module 🤞