@jonjohnsonjr 2023-09-20
> So I'd be fine with landing this as-is and trying to measure the impact of this on our mega module

SGTM -- collapsing at this level is nice because we get both fewer resources but also fewer HTTP requests to the registry (we do the whole `.att` tag at once), so this should make a nice dent.

Agree with adding a new resource if we want to collapse multiple archs/index into a single resource.