@mattmoor 2023-09-20
@jonjohnsonjr yeah, I also think it would be neat to consider having this "set" the attestations vs. augment them.  Or at least have an option to strip ~other attestations of the same predicate type, so it doesn't grow monotonically.